기능완성버전 -> 새로운 기능 or GUI 추가 필요

+++ cliBoard/src/com/board/view/View.java
@@ -1,52 +1,63 @@
 package com.board.view;
 
 import com.board.controller.PostController;
+import com.board.model.dao.Post;
 
+import java.io.File;
 import java.util.Scanner;
 
 public class View {
     private Scanner sc = new Scanner(System.in);
     private PostController pc = new PostController();
+
     public void mainView(){
 
-        while(true){
-            System.out.println("==================== 게시판 ====================");
-            System.out.print("글 번호\t\t");
-            System.out.println("글 제목");
+        File f = new File("adminInfo.txt");
+        // 관리자 정보 파일 있는지 확인
+        if( !f.exists() ) {
+            // 없으면 관리자 정보저장하기
+            new Post().saveAdminInfo();
 
-            try {
-                pc.postListPrint();
-            }catch (NullPointerException e){
-                // 처음 시작 시는 파일이 없으니까 Exception 방지
+        }else {
+            // 관리자 정보 있으면
+
+            while (true) {
+                System.out.println("==================== 게시판 ====================");
+                System.out.print("글 번호\t\t");
+                System.out.println("글 제목");
+
+                try {
+                    pc.postListPrint();
+                } catch (NullPointerException e) {
+                    // 처음 시작 시는 파일이 없으니까 Exception 방지
+                }
+
+                System.out.println();
+                System.out.println("================================================");
+                System.out.println("게시글을 보려면  \'post\'를 입력해주세요");
+                System.out.println("게시글을 작성하려면  \'write\'를 입력해주세요");
+                System.out.println("게시글을 삭제하려면  \'delete\'를 입력해주세요");
+                System.out.print("어떤 것을 하시겠습니까? : ");
+                String cho = sc.nextLine();
+
+                switch (cho.toLowerCase()) { // 입력문자 소문자로 통일
+                    case "post":
+                        viewPost();
+                        break;
+
+                    case "write":
+                        writePost();
+                        break;
+
+                    case "delete":
+                        deletePost();
+                        break;
+
+                    default:
+                        System.out.println("잘못입력하셨습니다. 다시 입력해주세요.");
+                        break;
+                }
             }
-
-
-            System.out.println();
-            System.out.println("================================================");
-            System.out.println("게시글을 보려면  \'post\'를 입력해주세요");
-            System.out.println("게시글을 작성하려면  \'write\'를 입력해주세요");
-            System.out.println("게시글을 삭제하려면  \'delete\'를 입력해주세요");
-            System.out.print("어떤 것을 하시겠습니까? : ");
-            String cho = sc.nextLine();
-
-            switch (cho.toLowerCase()){ // 입력문자 소문자로 통일
-                case "post" :
-                    viewPost();
-                    break;
-
-                case "write" :
-                    writePost();
-                    break;
-
-                case "delete" :
-                    deletePost();
-                    break;
-
-                default:
-                    System.out.println("잘못입력하셨습니다. 다시 입력해주세요.");
-                    break;
-            }
-
         }
     }
 
@@ -54,8 +65,11 @@ public class View {
         System.out.println("==================== 글 조회 ====================");
         System.out.print("게시글을 보려면 게시물 제목을 입력해주세요 : ");
         String postTitle = sc.nextLine();
+
+        System.out.print("글 내용 : ");
         pc.viewPost(postTitle);
 
+        System.out.println("=================================================");
 
     }
 
@@ -63,14 +77,14 @@ public class View {
         System.out.println("==================== 글 작성 ====================");
         System.out.print("게시글 제목을 입력해주세요 : ");
         String title = sc.nextLine();
+
         pc.createPost(title);
         // 입력받아서 게시글 생성.
         boolean stop = true;
 
         while(stop) {
 
-            System.out.println("게시글 내용을 입력해주세요 : ");
-            System.out.println("입력을 종료하려면 \' exit \'를 입력해주세요");
+            System.out.print("게시글 내용을 입력해주세요 : ");
             String content = sc.nextLine();
 
             if(content.toLowerCase().equals("exit")){
@@ -81,7 +95,11 @@ public class View {
                 // 내용 추가하기.
             }
 
+            System.out.println("입력을 종료하려면 \' exit \'를 입력해주세요");
+
+
         }
+        System.out.println("=================================================");
 
     }
 
@@ -121,6 +139,9 @@ public class View {
            System.out.println("메인으로 돌아갑니다.");
            return;
        }
+
+        System.out.println("=================================================");
+
     }
 
 }
